features(profilePageSetting)
set up profile page
[finish #167729791]
