+++ src/utils/helpers/pullErrors.js
@@ -10,6 +10,14 @@ const pullErrors = (errors) => {
   const email = [];
   const phone = [];
   const password = [];
+  const gender = [];
+  const birthday = [];
+  const preferredlanguage = [];
+  const currency = [];
+  const residentialaddress = [];
+  const role = [];
+  const department = [];
+  const linemanager = [];
   errors.forEach((error) => {
     const { msg } = error;
     const msgArray = msg.split(' ');
@@ -32,6 +40,30 @@ const pullErrors = (errors) => {
       case 'password':
         password.push(extractedError);
         break;
+      case 'gender':
+        gender.push(extractedError);
+        break;
+      case 'birthday':
+        birthday.push(extractedError);
+        break;
+      case 'preferredlanguage':
+        preferredlanguage.push(extractedError);
+        break;
+      case 'currency':
+        currency.push(extractedError);
+        break;
+      case 'residentialaddress':
+        residentialaddress.push(extractedError);
+        break;
+      case 'role':
+        role.push(extractedError);
+        break;
+      case 'department':
+        department.push(extractedError);
+        break;
+      case 'linemanager':
+        linemanager.push(extractedError);
+        break;
       default:
     }
   });
@@ -49,6 +81,30 @@ const pullErrors = (errors) => {
   }
   if (password.length > 0) {
     combinedErrors.password = password;
+  }
+  if (gender.length > 0) {
+    combinedErrors.gender = gender;
+  }
+  if (birthday.length > 0) {
+    combinedErrors.birthday = birthday;
+  }
+  if (preferredlanguage.length > 0) {
+    combinedErrors.preferrdlanguage = preferredlanguage;
+  }
+  if (currency.length > 0) {
+    combinedErrors.currency = currency;
+  }
+  if (residentialaddress.length > 0) {
+    combinedErrors.residentialaddress = residentialaddress;
+  }
+  if (role.length > 0) {
+    combinedErrors.role = role;
+  }
+  if (department.length > 0) {
+    combinedErrors.department = department;
+  }
+  if (linemanager.length > 0) {
+    combinedErrors.linemanager = linemanager;
   }
   return combinedErrors;
 };
